feat(s3): refactor storage interface and implement S3 client with event handling
